% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a_armata}
\alias{a_armata}
\title{Asparagopsis armata model parameters}
\format{
A named numeric vector with 34 elements:
\describe{
\item{\code{V_am}}{Maximum ammonium uptake rate (Michaelis-Menton uptake, mg gDW-1 d-1)}
\item{\code{K_am}}{Half-saturation constant for ammonium uptake (Michaelis-Menton uptake, mg m-3)}
\item{\code{M_am}}{Ammonium uptake rate slope (linear uptake, m3 gDW-1 d-1)}
\item{\code{C_am}}{Ammonium uptake rate constant (linear uptake, m3 gDW-1 d-1)}
\item{\code{V_ni}}{Maximum nitrate uptake rate (Michaelis-Menton uptake, mg gDW-1 d-1)}
\item{\code{K_ni}}{Half-saturation constant for nitrate uptake (Michaelis-Menton uptake, mg m-3)}
\item{\code{M_ni}}{Nitrate uptake rate slope (linear uptake, m3 gDW-1 d-1)}
\item{\code{C_ni}}{Nitrate uptake rate constant (linear uptake, m3 gDW-1 d-1)}
\item{\code{V_ot}}{Maximum other nitrogen uptake rate (Michaelis-Menton uptake, mg gDW-1 d-1)}
\item{\code{K_ot}}{Half-saturation constant for other nitrogen uptake (Michaelis-Menton uptake, mg m-3)}
\item{\code{M_ot}}{Other nitrogen uptake rate slope (linear uptake, m3 gDW-1 d-1)}
\item{\code{C_ot}}{Other nitrogen uptake rate constant (linear uptake, m3 gDW-1 d-1)}
\item{\code{Q_min}}{Minimum nitrogen quotient (mg gDW-1)}
\item{\code{Q_max}}{Maximum nitrogen quotient (mg gDW-1)}
\item{\code{K_c}}{Half-saturation constant for growth (mg gDW-1)}
\item{\code{mu}}{Maximum growth rate (d-1)}
\item{\code{D_m}}{Base loss rate (d-1)}
\item{\code{D_ve}}{Loss rate with velocity (m-1 s d-1)}
\item{\code{D_lo}}{Loss rate at low turbidity (d-1)}
\item{\code{D_mi}}{Loss rate at medium turbidity (d-1)}
\item{\code{D_hi}}{Loss rate at high turbidity (d-1)}
\item{\code{a_cs}}{Carbon-specific self-shading constant (m mg-1)}
\item{\code{I_o}}{Light saturation parameter (μmol photons m-2 s-1)}
\item{\code{T_opt}}{Optimal temperature (°C)}
\item{\code{T_min}}{Minimum temperature (°C)}
\item{\code{T_max}}{Maximum temperature (°C)}
\item{\code{S_opt}}{Optimal salinity (g L-1)}
\item{\code{S_min}}{Minimum salinity (g L-1)}
\item{\code{S_max}}{Maximum salinity (g L-1)}
\item{\code{h_a}}{Height-controlling parameter a (d-1)}
\item{\code{h_b}}{Height-controlling parameter b (d-1)}
\item{\code{h_c}}{Height-controlling parameter c (d-1)}
\item{\code{h_max}}{Maximum height (m)}
\item{\code{DWWW}}{Dry weight to wet weight ratio (gWW gDW-1)}
}
}
\source{
Parameters derived from literature values and used in \emph{Asparagopsis armata} growth modeling here (insert citation).
}
\usage{
a_armata
}
\description{
A named vector containing physiological and morphological parameters for the macroalgae species \emph{Asparagopsis armata} used in the growth model.
}
\examples{
# View all parameters
a_armata

# Access specific parameters
a_armata["mu"]  # Maximum growth rate
a_armata["h_max"]  # Maximum height
}
\seealso{
\code{\link[=grow_macroalgae]{grow_macroalgae()}} for the main growth model function.
}
\keyword{datasets}
