% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{abort_missing_parameter}
\alias{abort_missing_parameter}
\title{Custom error handling for missing parameters}
\usage{
abort_missing_parameter(param, place)
}
\arguments{
\item{param}{}

\item{place}{}
}
\description{
Custom error handling for missing parameters
}
