% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algae_height.R
\name{algae_height}
\alias{algae_height}
\title{Macroalgae height}
\usage{
algae_height(Nf, spec_params)
}
\arguments{
\item{Nf}{}

\item{spec_params}{A vector of named numbers. Must include the parameters:
\itemize{
\item \code{h_a}, \code{h_b} and \code{h_c}, parameters governing height change with \code{N_f}
\item \code{h_max}, maximum species height
}}
}
\value{
a scalar of macroalgae height (m)
}
\description{
Macroalgae height
}
\examples{
examples
}
