% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_Irradiance.R
\name{curve_Irradiance}
\alias{curve_Irradiance}
\title{Irradiance at site}
\usage{
curve_Irradiance(tspan = seq(1, 365, 1), a, b, c, period)
}
\arguments{
\item{tspan}{A vector of timesteps}

\item{a}{The mean irradience across timesteps at the site}

\item{b}{The temporal phase shift (the timestep where irradience is at maximum)}

\item{c}{The amplitude multiplier for the curve}

\item{period}{The total number of timesteps}
}
\value{
An irradiance vector for the given site
}
\description{
A short description...
}
\details{
A simple sine curve in the form:
\deqn{\begin{array}[ccc]
I = I_a + \frac{I_c}{period} \sin\left(\frac{2\pi(t+I_b) + \pi}{2}\right)
\end{array}}
}
\examples{
examples

}
