% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_Nutrient.R
\name{curve_Nutrient}
\alias{curve_Nutrient}
\title{Nutrient at site}
\usage{
curve_Nutrient(tspan = seq(1, 365, 1), a, b, c, period)
}
\arguments{
\item{tspan}{A vector of timesteps}

\item{a}{The mean nutrient concentration across timesteps at the site}

\item{b}{The temporal phase shift (the timestep where concentration is at maximum)}

\item{c}{The amplitude multiplier for the curve}

\item{period}{The total number of timesteps}
}
\value{
A vector of concentrations for the given site
}
\description{
A short description...
}
\details{
A simple sine curve in the form:
\deqn{\begin{array}[ccc]
N = N_a + \frac{N_c}{period} \sin\left(\frac{2\pi(t+N_b) + \pi}{2}\right)
\end{array}}
}
\examples{
examples

}
