% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_Temperature.R
\name{curve_Temperature}
\alias{curve_Temperature}
\title{Temperature at site}
\usage{
curve_Temperature(tspan = seq(1, 365, 1), a, b, c, period)
}
\arguments{
\item{tspan}{A vector of timesteps}

\item{a}{The mean temperature across timesteps at the site}

\item{b}{The temporal phase shift (the timestep where the temperature is at maximum)}

\item{c}{The amplitude multiplier for the curve}

\item{period}{The total number of timesteps}
}
\value{
A vector of temperatures for the given site
}
\description{
A short description...
}
\details{
A simple temperature curve in the form:
\deqn{\begin{array}[ccc]
T = T_a + \frac{T_c}{period}  * \sin\left(\frac{2\pi(t+T_b) + \pi}{2}\right)
\end{array}}
}
\examples{
examples
# Generate temperature fluctuations in a day
day <- seq(1, 24*60, 5)
temp <- temp_curve(tspan = day, 18.5, 14*60, 2.5)
plot(day, temp)

# Generate temperature fluctuations for part of the year
yearpart <- seq(90, 185, 1)
temp <- temp_curve(tspan = yearpart, 18.5, 60, 2.5, period = 365)
plot(yearpart, temp)

}
