% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic-functions.R
\name{hm_plot}
\alias{hm_plot}
\title{Plot height change over a range of fixed-N values}
\usage{
hm_plot(spec_params)
}
\arguments{
\item{spec_params}{A vector of named numbers. Must include the parameters:
\itemize{
\item \code{h_max}, maximum species height. Can be \code{NA}.
\item \code{h_a}, \code{h_b} and \code{h_c}, parameters governing height change with \code{N_f}. If not supplied algae height will always be \code{h_max}.
}}
}
\value{
a ggplot object of growth response over the specified range of fixed-N values
}
\description{
Plot height change over a range of fixed-N values
}
\examples{
examples
}
