% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_lim.R
\name{I_lim}
\alias{I_lim}
\title{Light limitation on growth}
\usage{
I_lim(Nf, I, kW, spec_params, site_params)
}
\arguments{
\item{Nf}{Fixed nitrogen (mg m\eqn{^{-3}})}

\item{I}{the surface irradiance, PAR (\eqn{\mu}mol photons m\eqn{^{-2}} s\eqn{^{-1}})}

\item{kW}{the light attenuation coefficient for open water (m\eqn{^{-1}})}

\item{spec_params}{a vector of named numbers. Must include:
\itemize{
\item \code{a_cs}, the carbon-specific self-shading constant
\item \code{I_o}, the light saturation parameter
\item \code{h_a}, \code{h_b} and \code{h_c}, parameters governing height change with \code{N_f}
\item \code{h_max}, maximum species height
}}

\item{site_params}{A vector of named numbers. Must include:
\itemize{
\item \code{d_top}, the below-surface  depth (m) of the top of the macroalgae culture
}}
}
\value{
a scalar of relative light limitation on growth (between 0 and 1)
}
\description{
Calculates the relative limitation on growth rate due to light availability, via:
\deqn{I_{lim} = \frac{e}{K \cdot d_{top}} \times \Biggl[e^{-\frac{I_z e^{-K \cdot d_{top}}}{I_o}} - e^{-\frac{I_z}{I_o}} \Biggr]}
where \eqn{I_{z}=I e^{-k_W \cdot d_{top}}} is the irradiance at the cultivation depth
and \eqn{K=k_{m}+kW} is the total attenuation coefficient.

\eqn{k_{m}} is the additional attenuation coefficient from macroalgae biomass, calculated as:
\deqn{k_{m} = a_{cs} \times N_f \times \text{max} \Biggl( \frac{h_{m}}{d_{top}}, 1 \Biggr) \times \frac{1}{\text{min}(h_{m}, d_{top})}}
where \eqn{h_m} is the macroalgae height.
}
\examples{
my_species <- c(I_o = 200, a_cs = 0.001, h_a = 750, h_b = 0.5, h_c = 0.01, h_max = 1)
my_site <- c(d_top = 1)
I <- seq(0, 2000, 10)
kW <- rnorm(length(I), 0.05, 0.01)

# Use purrr::map2 to calculate light limitation over a range of light and attenuation values, but fixed Nf
\dontrun{
df <- purrr::map2_dfr(I, kW, function(Iz, kWz) {
  data.frame(
    I = Iz,
    kW = kWz,
    lim = I_lim(Nf = 1000, I = Iz, kW = kWz, spec_params = my_species, site_params = my_site)
  )
})
}

}
\seealso{
\code{\link[=algae_height]{algae_height()}}
}
