% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic-functions.R
\name{I_plot}
\alias{I_plot}
\title{Plot of species growth response over a range of surface irradience values}
\usage{
I_plot(Nf, spec_params, site_params, irr = seq(0, 2500, 10))
}
\arguments{
\item{Nf}{Fixed nitrogen (mg m\eqn{^{-3}})}

\item{spec_params}{a vector of named numbers. Must include:
\itemize{
\item \code{a_cs}, the carbon-specific self-shading constant
\item \code{I_o}, the light saturation parameter
\item \code{h_a}, \code{h_b} and \code{h_c}, parameters governing height change with \code{N_f}
\item \code{h_max}, maximum species height
}}

\item{site_params}{A vector of named numbers. Must include:
\itemize{
\item \eqn{d_{top}} the below-surface  depth (m) of the top of the macroalgae culture
}}

\item{irr}{a vector of the range of surface irradiance values to be tested (defaults to seq(0, 2500, 10))}
}
\value{
a ggplot object of growth response over the specified range of surface irradience values
}
\description{
Plot of species growth response over a range of surface irradience values
}
\examples{
my_species <- c(I_o = 150, a_cs = 0.005, h_max = 1)
my_site <- c(kW = 0.6, d_top = 1)
I_plot(Nf = 150, spec_params = my_species, site_params = my_site)

}
\seealso{
\code{\link[=I_lim() algae_height]{I_lim() algae_height()}}
}
