% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_uptake.R
\name{lin_uptake}
\alias{lin_uptake}
\title{Uptake rate (linear)}
\usage{
lin_uptake(conc, M, C)
}
\arguments{
\item{conc}{external substrate concentration}

\item{M}{the slope of N uptake with increasing substrate concentration}

\item{C}{the intercept}
}
\value{
the rate of uptake at the specified external concentration
}
\description{
Uptake rate (linear)
}
\examples{
examples
}
