% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{loss}
\alias{loss}
\title{Loss through fragmentation}
\usage{
loss(U0 = 0, turbulence = NA, spec_params)
}
\arguments{
\item{U0}{incoming incident water velocity (m/s)}

\item{turbulence}{optional level of turbulence. Options are "none" (or "static"), "low", "medium" (or "mid"), and "high".}

\item{spec_params}{a vector of named numbers. All default to 0 if not supplied. Can include:
\itemize{
\item \code{D_m}, a constant (base) loss rate
\item \code{D_ve}, the linear coefficient of loss with laminar velocity
\item \code{D_st}, the loss rate in static water (turbulence = "none")
\item \code{D_lo}, the loss rate in turbulent water (turbulence = "low")
\item \code{D_mi}, the loss rate in turbulent water (turbulence = "medium")
\item \code{D_hi}, the loss rate in turbulent water (turbulence = "high")
}}
}
\value{
a scalar for macroalgae loss (\% d-1)
}
\description{
The macroalgae fragmentation rate is calculated as:
\begin{equation}
D_m = D_{velocity} \times U_0 + D_{turbulence} + D_{base}
\end{equation}
where $D_{velocity}$ is the linear coefficient of loss with laminar velocity,
$D_{turbulence}$ is the loss rate at the specified level of turbulence (static, low, medium, or high), and
$D_{base}$ is the base loss rate.
}
\examples{
examples

}
