% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroalgae-state-functions.R
\name{MM_uptake}
\alias{MM_uptake}
\title{Uptake rate (Michaelis-Menton)}
\usage{
MM_uptake(conc, V, K)
}
\arguments{
\item{conc}{external substrate concentration}

\item{V}{the maximum uptake rate \eqn{V_{max}}}

\item{K}{the half-saturation constant \eqn{K_{c}}}
}
\value{
the rate of uptake at the specified external concentration
}
\description{
Uptake rate (Michaelis-Menton)
}
\examples{
examples
}
