% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic-functions.R
\name{N_plot}
\alias{N_plot}
\title{Plot a species uptake rate over a range of N concentrations}
\usage{
N_plot(
  spec_params,
  uptake_shape,
  abbreviations = c(ni = "nitrate", am = "ammonium", ot = "other")
)
}
\arguments{
\item{spec_params}{a vector of named numbers. Must include:
\itemize{
\item \code{M} and \code{C} for linear uptake, OR
\item \code{V} and \code{K} for Michaelis-Menton uptake
\item \code{uptake_shape} of "linear" or "MM" (or "Michaelis-Menton"). Not required if only one set of the above parameters are provided.
}}

\item{uptake_shape}{kinetic shape for substrate uptake. One of "linear" or "Michaelis-Menton" (or "MM"). Defaults to Michaelis-Menton.}

\item{conc}{a vector of the range of N concentrations to be tested (defaults to seq(0, 50, 0.25))}
}
\value{
a ggplot object of nitrogen uptake rates
}
\description{
Plot a species uptake rate over a range of N concentrations
}
\examples{
examples
}
