% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_rel.R
\name{N_rel}
\alias{N_rel}
\title{Relative internal nitrogen quotient}
\usage{
N_rel(
  Nf = NULL,
  Ns = NULL,
  Q_int = NULL,
  N_int = NULL,
  Q_rel = NULL,
  N_rel = NULL,
  spec_params
)
}
\arguments{
\item{N_int}{the internal nitrogen concentration (percentage, as decimal)}

\item{spec_params}{a vector of named numbers. Must include:
\itemize{
\item \code{N_min}, the minimum internal nitrogen concentration
\item \code{N_max}, the maximum internal nitrogen concentration
}}
}
\value{
relative internal nitrogen quotient
}
\description{
Relative internal nitrogen quotient
}
\seealso{
\code{\link[=N_int(), Q_rel(), Q_int]{N_int(), Q_rel(), Q_int()}}
}
