% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_Nf_conversions.R
\name{Nf_to_biomass}
\alias{Nf_to_biomass}
\title{Convert Nf to biomass}
\usage{
Nf_to_biomass(Nf, Ns, Q_int = NULL, Q_rel = 0.5, spec_params, dry = T)
}
\arguments{
\item{Nf}{Fixed nitrogen (mg m\eqn{^{-3}})}

\item{Ns}{Stored nitrogen (mg m\eqn{^{-3}})}

\item{Q_int}{Non-dimensionalised internal nutrient quotient}

\item{spec_params}{A vector of named numbers. Must include:
\itemize{
\item \code{Q_min}, the minimum internal nutrient quotient
}}

\item{dry}{logical, return dry or wet biomass. If dry = F, \code{spec_params['DWWW']} must be provided}
}
\value{
dry (or wet) biomass, mg m-3
}
\description{
Convert Nf to biomass
}
\details{
Not all parameters need to be provided to this function.
}
\examples{
my_species <- c(DWWW = 7.5, Q_min = 20, Q_max = 45)
starting_Nf <- 150 # mg m-3
starting_Ns <- 50 # mg m-3

# Using default Q_rel = 0.5
Nf_to_biomass(Nf = starting_Nf, Ns = starting_Ns, spec_params = my_species, dry = T)

# Using a specific Q_int
Nf_to_biomass(biomass = starting_biomass, Q_int = 30, spec_params = my_species, dry = T)

}
\seealso{
\code{\link[=biomass_to_Nf]{biomass_to_Nf()}}, \code{\link[=Q_rel]{Q_rel()}}, \code{\link[=Q_int]{Q_int()}}
}
