% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroalgae-state-functions.R
\name{NN}
\alias{NN}
\title{Relative internal nitrogen quotient}
\usage{
NN(N_int, spec_params)
}
\arguments{
\item{N_int}{the internal nitrogen concentration (percentage, as decimal)}

\item{spec_params}{A vector of named numbers. Must include:
\itemize{
\item \eqn{N_{min}}, the minimum internal nitrogen concentration
\item \eqn{N_{max}}, the maximum internal nitrogen concentration
}}
}
\value{
relative internal nitrogen quotient
}
\description{
Relative internal nitrogen quotient
}
