% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic-functions.R
\name{plot_h}
\alias{plot_h}
\title{Plot height change over a range of fixed-N values}
\usage{
plot_h(Nf = seq(1, 4500, 1), spec_params)
}
\arguments{
\item{Nf}{}

\item{spec_params}{}
}
\value{
a ggplot object of growth response over the specified range of fixed-N values
}
\description{
Plot height change over a range of fixed-N values
}
\examples{
examples
}
