% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic-functions.R
\name{plot_I_range}
\alias{plot_I_range}
\title{Plot of species growth response over a range of surface irradience values}
\usage{
plot_I_range(
  I_range = seq(1, 2500, 1),
  I_top = c(1, 2, 5),
  spec_params,
  site_params
)
}
\arguments{
\item{I_range}{}

\item{I_top}{}

\item{kW}{}

\item{QQ}{}

\item{nf}{}
}
\value{
a ggplot object of growth response over the specified range of surface irradience values
}
\description{
Plot of species growth response over a range of surface irradience values
}
\examples{
examples
}
\seealso{
\code{\link[=I_lim() algae_height]{I_lim() algae_height()}}
}
