% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroalgae-state-functions.R
\name{Q}
\alias{Q}
\title{Non-dimentionalised internal nutrient quotient}
\usage{
Q(Nf, Ns, spec_params)
}
\arguments{
\item{Nf}{Fixed nitrogen (mg m\eqn{^{-3}})}

\item{Ns}{Stored nitrogen (mg m\eqn{^{-3}})}

\item{spec_params}{A vector of named numbers. Must include:
\itemize{
\item \eqn{Q_{min}}, the minimum internal nutrient quotient
}}
}
\value{
the non-dimensionalised internal nutrient quotient
}
\description{
Calculates the internal nutrient quotient \eqn{Q} via:
\deqn{\begin{array}[ccc] 
Q &=& Q_{min} \left(1 + \frac{N_s}{N_f}\right)
\end{array}}
}
\examples{
examples
}
