% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_lim.R
\name{Q_lim}
\alias{Q_lim}
\title{Q_int limitation on growth}
\usage{
Q_lim(Nf, Ns, spec_params)
}
\arguments{
\item{Nf}{Fixed nitrogen (mg m\eqn{^{-3}})}

\item{Ns}{Stored nitrogen (mg m\eqn{^{-3}})}

\item{spec_params}{a vector of named numbers. Must include:
\itemize{
\item \code{Q_min}, the minimum nutrient quotient
\item \code{Q_max}, the maximum nutrient quotient
\item \code{K_c}, the half-saturation growth constant
}}
}
\value{
a scalar of relative limitation from internal nutrient reserves on growth (between 0 and 1)
}
\description{
Calculates the internal nutrient quotient \eqn{Q_{int}} and its relative effect on growth rate \eqn{Q_{lim}} via:
\deqn{\begin{array}[ccc] 
Q_{int} &=& Q_{min} \left(1 + \frac{N_s}{N_f}\right) \\
Q_{lim} &=& \frac{Q_{int} - Q_{min}}{Q_{int} - K_c}
\end{array}}
}
\examples{
examples
}
\seealso{
\code{\link[=Q_int]{Q_int()}}, \code{\link[=Q_rel]{Q_rel()}}
}
