% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_lim.R
\name{S_lim}
\alias{S_lim}
\title{Salinity limitation on growth}
\usage{
S_lim(Sal, spec_params)
}
\arguments{
\item{Sal}{salinity to evaluate (g/L)}

\item{spec_params}{a vector of named numbers. Must include:
\itemize{
\item \code{S_opt} the optimum salinity for macroalgae growth
\item \code{S_min} the minimum salinity for macroalgae growth (when \code{Sal} <= \code{S_min}, growth = 0)
\item \code{S_max} the maximum salinity for macroalgae growth (when \code{Sal} >= \code{S_max}, growth = 0)
}}
}
\value{
a scalar of relative salinity limitation on growth (between 0 and 1)
}
\description{
Given species parameters, returns the relative limitation on growth rate according to a CTMI curve:
\deqn{
     S_{lim} &=& \frac{(Sal-S_{max})(Sal-S_{min})^2}{(S_{opt}-S_{min})[(S_{opt}-S_{min})(Sal-S_{opt})-(S_{opt}-S_{max})(S_{opt}+S_{min}-2Sal)]}
}
The CTMI curve was formulated for temperature.
}
\examples{
my_seaweed <- c(S_opt = 20, S_min = 5, S_max = 30)

S_lim(Sal = 22, spec_params = my_seaweed)

S_range <- 1:30
sapply(S_range, S_lim, spec_params = my_seaweed)

}
