% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Secchi_Kd.R
\name{Secchi_from_Kd}
\alias{Secchi_from_Kd}
\title{Secchi depth from Kd}
\usage{
Secchi_from_Kd(Kd)
}
\arguments{
\item{Kd}{attenuation parameter for PAR}
}
\value{
Secchi disk depth (m)
}
\description{
Calculates the light attenuation coefficient in water via:
\deqn{\begin{array}[ccc] 
K_d(PAR) &=& 1.584 × SDD^{-0.894}
\end{array}}
where SDD is the Secchi disk depth. Equation is adapted from Zhang, Y., Liu, X., Yin, Y., Wang, M., & Qin, B. (2012). Predicting the light attenuation coefficient through Secchi disk depth and beam attenuation coefficient in a large, shallow, freshwater lake. Hydrobiologia, 693(1), 29–37. https://doi.org/10.1007/s10750-012-1084-2.
}
