% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth-limiting-functions.R
\name{T_lim}
\alias{T_lim}
\title{Temperature limitation on growth}
\usage{
T_lim(TT, spec_params)
}
\arguments{
\item{TT}{temperature to evaluate}

\item{spec_params}{a vector of named numbers. Must include:
\itemize{
\item \eqn{T_{opt}} the optimum temperature for macroalgae growth
\item \eqn{T_{min}} the minimum temperature for macroalgae growth (when \eqn{T < T_{min}}, growth = 0)
\item \eqn{T_{max}} the maximum temperature for macroalgae growth (when \eqn{T > T_{max}}, growth = 0)
}}
}
\value{
a scalar of relative temperature limitation on growth (between 0 and 1)
}
\description{
Given species parameters returns the relative limitation on growth rate according to a CTMI curve.
}
\examples{
examples examples

}
