% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic-functions.R
\name{T_plot}
\alias{T_plot}
\title{Plot of species growth response over a range of temperatures}
\usage{
T_plot(spec_params, temp = seq(-10, 45, 0.25))
}
\arguments{
\item{spec_params}{a vector of named numbers. Must include:
\itemize{
\item \code{T_opt} the optimum temperature for macroalgae growth
\item \code{T_min} the minimum temperature for macroalgae growth (when \code{T_c} < \code{T_min}, growth = 0)
\item \code{T_max} the maximum temperature for macroalgae growth (when \code{T_c} > \code{T_max}, growth = 0)
}}

\item{temp}{a vector of the range of temperatures to be tested (defaults to seq(-10, 45, 0.25))}
}
\value{
a ggplot object of growth response over the specified temperature range
}
\description{
A shortcut function to quickly see a species' response to a range of temperatures.
}
\examples{
spec_params <- c(T_opt = 25, T_min = 5, T_max = 35)
T_plot(spec_params)

}
\seealso{
\code{\link[=T_lim]{T_lim()}}
}
