% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drag_functions.R
\name{u_c}
\alias{u_c}
\title{Relative water attenuation within canopy}
\usage{
u_c(
  U0,
  macro_state,
  site_params,
  spec_params,
  constants = c(s = 0.0045, gam = 1.13, a2 = 0.2^2, Cb = 0.0025)
)
}
\arguments{
\item{U0}{incoming incident water velocity (m/s)}

\item{macro_state}{vector of named numbers. Must include:
\itemize{
\item \code{biomass}, macroalgae wet weight (g)
\item \code{hm}, algae height (m)
}}

\item{site_params}{vector of named numbers. Must include:
\itemize{
\item \code{hz}, total water depth (m)
\item \code{hc}, vertical water column occupied by the canopy (m)
\item \code{d_top}, depth of the top of the canopy beneath the water surface (m)
}}

\item{spec_params}{vector of named numbers. Must include:
\itemize{
\item \code{SA_WW}, conversion of wet weight to surface area
}}

\item{constants}{vector of named numbers defining extra constants for the attenuation submodel. Must include all constants, which have the following default values:
\itemize{
\item \code{s} = 0.0045
\item \code{gam} = 1.13
\item \code{a2} = 0.2^2
\item \code{Cb} = 0.0025
}}
}
\value{
the relative water attenuation coefficient (u_c)
}
\description{
Calculates relative water attenuation within the canopy, based on Plew, D. R. (2011). "Depth-Averaged Drag Coefficient for Modeling Flow through Suspended Canopies". Journal of Hydraulic Engineering, 137(2), 234–247.
}
\seealso{
\code{\link[=height]{height()}}, \code{\link[=u_b]{u_b()}}, \code{\link[=C_t]{C_t()}}
}
