% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algae_height.R
\name{algae_height}
\alias{algae_height}
\title{Macroalgae height}
\usage{
algae_height(Nf, spec_params)
}
\arguments{
\item{Nf}{}

\item{spec_params}{A vector of named numbers. Must include the parameters:
\itemize{
\item \code{h_max}, maximum species height. If only \code{h_max} is supplied then algae height = \code{h_max}
\item \code{h_a}, \code{h_b} and \code{h_c}, parameters governing height change with \code{N_f}
}}
}
\value{
a scalar of macroalgae height (m)
}
\description{
Macroalgae height
}
\details{
\code{h_a}, \code{h_b} and \code{h_c} default to 1000, 1 and 0 if not supplied. Algae height therefore defaults to \code{Nf/1000} if no parameters are supplied.
}
\examples{
examples
}
