% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algae_height.R
\name{algae_height}
\alias{algae_height}
\title{Macroalgae height}
\usage{
algae_height(Nf, spec_params)
}
\arguments{
\item{Nf}{}

\item{spec_params}{A vector of named numbers. Must include the parameters:
\itemize{
\item \code{h_max}, maximum species height. If only \code{h_max} is supplied then algae height == \code{h_max}
\item \code{h_a}, \code{h_b} and \code{h_c}, parameters governing height change with \code{N_f}
}}
}
\value{
a scalar of macroalgae height (m)
}
\description{
Macroalgae height
}
\details{
Calculates macroalgae height as:
\deqn{h_m = \left(\frac{N_f}{h_a}\right)^{h_b} + h_c}
up to a maximum of \verb{h_\{max\}}.

Defaults are \eqn{h_a=1000}, \eqn{h_b=1} and \eqn{h_c=0}.
Algae height therefore defaults to \eqn{N_f \times 10^{-3}} if no parameters are supplied.
}
\examples{
Examples TBD
}
