% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroalgae-state-functions.R
\name{algae_height}
\alias{algae_height}
\title{Macroalgae height}
\usage{
algae_height(Nf, spec_params)
}
\arguments{
\item{Nf}{}

\item{spec_params}{A vector of named numbers. Must include the parameters:
\itemize{
\item \eqn{h_a}, \eqn{h_b} and \eqn{h_c}, parameters governing height change with \eqn{N_f}
}}
}
\value{
a scalar of macroalgae height (m)
}
\description{
Macroalgae height
}
\examples{
examples
}
