% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_to_Nf.R
\name{biomass_to_Nf}
\alias{biomass_to_Nf}
\title{Convert biomass to Nf (and Ns)}
\usage{
biomass_to_Nf(biomass, Q_int = NULL, Q_rel = 0.5, spec_params, dry = T)
}
\arguments{
\item{biomass}{wet (or dry) biomass, mg m-3}

\item{Q_int}{Non-dimensionalised internal nutrient quotient}

\item{Q_rel}{the non-dimensionalised relative internal nutrient quotient (\eqn{Q_{rel}})}

\item{spec_params}{a vector of named numbers. Must include:
\itemize{
\item \code{DWWW} (if dry = F), the conversion from dry weight to wet weight
\item \code{Q_min}, the minimum internal nutrient quotient
\item \code{Q_max}, the maximum internal nutrient quotient
}}

\item{dry}{logical, return dry or wet biomass. If dry = F, spec_params\link{'DWWW'} must be provided}
}
\value{
Nf, mg m-3
}
\description{
Convert biomass to Nf (and Ns)
}
\examples{
examples
}
\seealso{
\code{\link[=Nf_to_biomass(), Q_rel(), Q_int]{Nf_to_biomass(), Q_rel(), Q_int()}}
}
