% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_to_Nf.R
\name{biomass_to_Nf}
\alias{biomass_to_Nf}
\title{Convert biomass to Nf (and Ns)}
\usage{
biomass_to_Nf(biomass, N_int, spec_params, dry = T)
}
\arguments{
\item{biomass}{wet (or dry) biomass, mg m-3}

\item{N_int}{the internal nitrogen concentration (percentage, as decimal)}

\item{spec_params}{a vector of named numbers. Must include:
\itemize{
\item \code{N_min}, the minimum internal nitrogen concentration
\item \code{N_max}, the maximum internal nitrogen concentration
}}

\item{dry}{logical, return dry or wet biomass. If dry = F, spec_params\link{'DWWW'} must be provided}
}
\value{
a vector of c(Nf, Ns)
}
\description{
Convert biomass to Nf (and Ns)
}
\examples{
examples
}
\seealso{
\link{N_int(), Nf_to_biomass}
}
