% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_Nf_conversions.R
\name{biomass_to_Nf}
\alias{biomass_to_Nf}
\title{Convert biomass to Nf and Ns}
\usage{
biomass_to_Nf(biomass, Q_int = NULL, Q_rel = 0.5, spec_params, dry = T)
}
\arguments{
\item{biomass}{whether dry (default) or wet biomass is provided, mg m-3}

\item{Q_int}{Non-dimensionalised internal nutrient quotient}

\item{Q_rel}{the non-dimensionalised relative internal nutrient quotient (\eqn{Q_{rel}}). If neither of Q_int and Q_rel are provided the default Q_rel of 0.5 will be used.}

\item{spec_params}{a vector of named numbers. Must include:
\itemize{
\item \code{DWWW} (if dry = F), the conversion from dry weight to wet weight
\item \code{Q_min}, the minimum internal nutrient quotient (mg gDW-1)
\item \code{Q_max}, the maximum internal nutrient quotient
}}

\item{dry}{logical, return dry or wet biomass. If dry = F, spec_params\link{'DWWW'} must be provided}
}
\value{
Nf, mg m-3
}
\description{
B = (N_f + N_s)/Q_min
Q_rel = Q_min * (1 + N_s/N_f)
}
\examples{
Examples TBD
}
\seealso{
\code{\link[=Nf_to_biomass(), Q_rel(), Q_int]{Nf_to_biomass(), Q_rel(), Q_int()}}
}
