% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_t.R
\name{C_t}
\alias{C_t}
\title{Total drag coefficient}
\usage{
C_t(
  u_c,
  u_b,
  macro_state = c(biomass, hm),
  SA_WW = 0.5 * (0.0306/2),
  site_params,
  constants = c(s = 0.0045, gam = 1.13, a2 = 0.2^2, Cb = 0.0025)
)
}
\arguments{
\item{macro_state}{vector of named numbers. Must include:
\itemize{
\item \code{biomass}, macroalgae wet weight (g)
\item \code{hm}, algae height (m)
}}

\item{site_params}{vector of named numbers. Must include:
\itemize{
\item \code{hz}, total water depth (m)
\item \code{hc}, vertical water column occupied by the canopy (m)
\item \code{d_top}, depth of the top of the canopy beneath the water surface (m)
}}

\item{constants}{vector of named numbers defining extra constants for the attenuation submodel. Must include:
\itemize{
\item \code{s} = 0.0045
\item \code{gam} = 1.13
\item \code{a2} = 0.2^2
\item \code{Cb} = 0.0025
}}
}
\value{
The total drag coefficient C_t
}
\description{
Total drag coefficient
}
