% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_grow.R
\name{check_grow}
\alias{check_grow}
\title{Pre-check all parameters for grow_macroalgae()}
\usage{
check_grow(
  t = 1:30,
  temperature,
  salinity,
  light,
  kW,
  velocity,
  nitrate,
  ammonium,
  ni_uptake,
  am_uptake,
  site_params,
  spec_params,
  initials,
  sparse_output = T,
  other_constants = c(s = 0.0045, gam = 1.13, a2 = 0.2^2, Cb = 0.0025)
)
}
\arguments{
\item{temperature}{a vector of daily temperatures (\eqn{^{\circ}}C)}

\item{salinity}{a vector of daily salt concentrations (g L\eqn{^{-1}})}

\item{light}{a vector of surface light (umol m\eqn{^{-2}} s\eqn{^{-1}})}

\item{kW}{a vector of light attenuation coefficients for open water (m\eqn{^{-1}})}

\item{velocity}{a vector of water velocities (m s\eqn{^{-1}})}

\item{nitrate}{a vector of nitrate concentrations (mg m\eqn{^{-3}})}

\item{ammonium}{a vector of ammonium concentrations (mg m\eqn{^{-3}})}

\item{ni_uptake}{shape for nitrate uptake, either "MM" for Michaelis-Menton saturation kinetics or "linear" for linear uptake}

\item{am_uptake}{shape for ammonium uptake, either "MM" for Michaelis-Menton saturation kinetics or "linear" for linear uptake}

\item{site_params}{a named vector of species-specific parameters - see details}

\item{spec_params}{a named vector of site-specific parameters - see details}

\item{initials}{a named vector of the macroalgae starting conditions}

\item{sparse_output}{logical, whether to include input vectors and other non-essential information in final dataframe (default = TRUE)}

\item{other_constants}{a named vector of other constants for intermediate equations (see \code{u_c()})}
}
\value{
printout of potential errors for main function
}
\description{
Check that all parameters, inputs and settings are correct for the \code{grow_macroalgae()} function. Gives a report on what needs to be modified for the main function to run smoothly. This is to avoid the main function slowing down to give endless warnings and messages.
}
\details{
Example csv with all the spec_params & site_params required?
}
\examples{
"see here" link?
}
