% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_grow.R
\name{check_grow}
\alias{check_grow}
\title{Pre-check all parameters for grow_macroalgae()}
\usage{
check_grow(
  start,
  grow_days,
  temperature,
  salinity,
  light,
  velocity,
  nitrate,
  ammonium,
  other_N = NA,
  ni_uptake,
  am_uptake,
  ot_uptake = NA,
  site_params,
  spec_params,
  initials,
  sparse_output = T,
  track_limiting = T,
  other_constants = c(s = 0.0045, gam = 1.13, a2 = 0.2^2, Cb = 0.0025)
)
}
\arguments{
\item{start}{numeric, start of the growth period (day of at-sea deployment). Defaults to 1.}

\item{grow_days}{integer, number of day in growing period - if missing will take the length of the temperature vector}

\item{temperature}{a vector of daily temperatures (C)}

\item{salinity}{a vector of daily salt concentrations (g L-1)}

\item{light}{a vector of surface light (umol m-2 s-1)}

\item{velocity}{a vector of water velocities (m s-1)}

\item{nitrate}{a vector of nitrate concentrations (mg m-3)}

\item{ammonium}{a vector of ammonium concentrations (mg m-3)}

\item{other_N}{a vector of other nitrogen concentrations (mg N m-3) - NOT IN USE}

\item{site_params}{a named vector of species-specific parameters - see details}

\item{spec_params}{a named vector of site-specific parameters - see details}

\item{initials}{a named vector of the macroalgae starting conditions}

\item{sparse_output}{logical, whether to include input vectors and other non-essential information in final dataframe (default = TRUE)}

\item{track_limiting}{logical, whether to track a single "limiting" factor (see details)}

\item{other_constants}{a named vector of miscellaneous constants (see u_c)}

\item{kd}{a vector of light attenuation parameters (m-1)}
}
\value{
printout of potential errors with main function
}
\description{
Check that all parameters, inputs and settings are correct for the \code{grow_macroalgae()} function. Gives a report on what needs to be included. This is to avoid the main function slowing down to give endless warnings and messages.
}
\details{
Example csv with all the spec_params & site_params required?
}
\examples{
"see here" link?
}
