% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_limiting.R
\name{get_limiting}
\alias{get_limiting}
\title{Determine limiting factor in growth}
\usage{
get_limiting(output, spec_params)
}
\arguments{
\item{output}{matrix, direct output of 'grow_macroalgae' function (can be sparse)}

\item{spec_params}{a named vector of site-specific parameters - same as input to 'grow_macroalgae' function}
}
\value{
vector of limiting factors at each timestep
}
\description{
For each timestep in the growth output matrix, determine the factor most likely to be limiting growth.
}
\details{
This function gives an idea of what factor(s) limit growth at each timestep. The choices (in order) are:
\enumerate{
\item T_lim
\item I_lim
\item S_lim
\item Q_lim - internal nutrient quotient
\item Ns_to_Nf - not enough stored nutrients (Ns) were available to be fixed at maximum rate
\item Nf_loss - more Nf was lost than fixed
\item conc_nitrate/conc_ammonium/conc_other - there were not enough ambient nutrients to be taken up at the maximum rate
\item Ns_fixed_not_replaced - the Ns fixed last timestep was not replaced this timestep (Q_rel is declining)
\item Ns_loss_not_replaced - the Ns lost last timestep was not replaced this timestep (Q_rel is declining)
}
}
\seealso{
\code{\link[=grow_macroalgae]{grow_macroalgae()}}
}
