% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uptake.R
\name{get_uptake}
\alias{get_uptake}
\title{Get the correct uptake rate}
\usage{
get_uptake(conc, uptake_shape, Nform_abbr, spec_params)
}
\arguments{
\item{conc}{}

\item{uptake_shape}{}

\item{Nform_abbr}{}

\item{spec_params}{}
}
\description{
Get the correct uptake rate
}
