% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uptake.R
\name{get_uptake}
\alias{get_uptake}
\title{Get the correct uptake rate}
\usage{
get_uptake(conc, uptake_shape = NA, Nform_abbr, spec_params)
}
\arguments{
\item{conc}{substrate concentration}

\item{uptake_shape}{kinetic shape for substrate uptake. One of "linear" or "Michaelis-Menton" (or "MM"). Defaults to Michaelis-Menton.}

\item{Nform_abbr}{the abbreviation used in spec_params for the relevant nitrogen form. See details.}

\item{spec_params}{a vector of named numbers. Must include:
\itemize{
\item \code{M} and \code{C} for linear uptake, OR
\item \code{V} and \code{K} for Michaelis-Menton uptake
\item \code{uptake_shape} of "linear" or "MM" (or "Michaelis-Menton"). Not required if only one set of the above parameters are provided.
}}
}
\description{
Get the correct uptake rate
}
\details{
Gives the substrate uptake rate based on the shape and species-specific parameters supplied.
If no shape is supplied, attempts to infer the correct shape from supplied parameters.

\code{Nform_abbr} allows the function to link species parameters with concentration.
E.g. if \code{Nform_abbr} = "amm" (for ammonium) the function will look for \code{M_amm} and \code{C_amm} or \code{V_amm} and \code{K_amm} in spec_params and will ignore other uptake parameters which may be included for other substrates.
}
\seealso{
\code{\link[=lin_uptake(), MM_uptake]{lin_uptake(), MM_uptake()}}
}
