% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN-model-function.R
\name{grow_macroalgae}
\alias{grow_macroalgae}
\title{Grow macroalgae}
\usage{
grow_macroalgae(
  start,
  grow_days,
  temperature,
  light,
  velocity,
  nitrate,
  ammonium,
  other_N,
  ni_uptake,
  am_uptake,
  ot_uptake,
  site_params,
  spec_params,
  other_constants = c(rL = 0.2, Rd = 0.1),
  initials
)
}
\arguments{
\item{start}{}

\item{grow_days}{}

\item{temperature}{}

\item{light}{}

\item{velocity}{}

\item{nitrate}{}

\item{ammonium}{}

\item{other_N}{}

\item{site_params}{}

\item{spec_params}{see details}

\item{other_constants}{}

\item{initials}{}
}
\value{
dataframe of outputs
}
\description{
The main deal
}
\details{
Example csv with all the spec_params required?
}
\examples{
"see here" link?
}
