% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{loss}
\alias{loss}
\title{Loss through fragmentation}
\usage{
loss(U, turbulence = 0, spec_params)
}
\arguments{
\item{turbulence}{optional level of turbulence. Options are "none"/"static", "low", "medium"/"mid", and "high".}

\item{spec_params}{a vector of named numbers. All default to 0 if not supplied. Can include:
\itemize{
\item \code{D_m}, a constant (base) loss rate
\item \code{D_ve}, the linear coefficient of loss with laminar velocity
\item \code{D_st}, the loss rate in static water (turbulence = 0, "none" or "static")
\item \code{D_lo}, the loss rate in turbulent water (turbulence = 1 or "low")
\item \code{D_mi}, the loss rate in turbulent water (turbulence = 2 or "medium")
\item \code{D_hi}, the loss rate in turbulent water (turbulence = 3 or "high")
}}
}
\value{
a scalar for macroalgae loss (\% d-1)
}
\description{
The macroalgae fragmentation rate ($D$) is calculated as:
\deqn{D = D_{ve} \times U_0 + D_{tu} + D_{m}}
where \eqn{D_{ve}} is the linear coefficient of loss with laminar velocity (\equ{U_0}),
\eqn{D_{tu}} is the loss rate at the specified level of turbulence (static, low, medium, or high), and
\eqn{D_{m}} is the base loss rate.
}
\details{
Note that parameter \code{U0} has no default.
If one is not supplied, loss due to velocity will default to 0 regardless of the value of \code{D_ve} supplied.
}
\examples{
examples

}
