% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nf_to_biomass.R
\name{Nf_to_biomass}
\alias{Nf_to_biomass}
\title{Convert Nf to biomass}
\usage{
Nf_to_biomass(Nf, Ns, N_int, dry = T, ...)
}
\arguments{
\item{Nf}{Fixed nitrogen (mg m\eqn{^{-3}})}

\item{Ns}{Stored nitrogen (mg m\eqn{^{-3}})}

\item{N_int}{the internal nitrogen concentration (percentage, as decimal)}

\item{dry}{logical, return dry or wet biomass. If dry = F, spec_params\link{'DWWW'} must be provided}
}
\value{
dry (or wet) biomass, mg m-3
}
\description{
Convert Nf to biomass
}
\examples{
examples
}
\seealso{
\link{N_int(), biomass_to_Nf}
}
