% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_Nf_conversions.R
\name{Nf_to_biomass}
\alias{Nf_to_biomass}
\title{Convert Nf to biomass}
\usage{
Nf_to_biomass(Nf, Ns, Q_int = NULL, Q_rel = 0.5, spec_params, dry = T)
}
\arguments{
\item{Nf}{Fixed nitrogen (mg m\eqn{^{-3}})}

\item{Ns}{Stored nitrogen (mg m\eqn{^{-3}})}

\item{Q_int}{Non-dimensionalised internal nutrient quotient}

\item{spec_params}{A vector of named numbers. Must include:
\itemize{
\item \code{Q_min}, the minimum internal nutrient quotient
}}

\item{dry}{logical, return dry or wet biomass. If dry = F, spec_params\link{'DWWW'} must be provided}
}
\value{
dry (or wet) biomass, mg m-3
}
\description{
Convert Nf to biomass
}
\examples{
examples
}
\seealso{
\code{\link[=biomass_to_Nf(), Q_rel(), Q_int]{biomass_to_Nf(), Q_rel(), Q_int()}}
}
