% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic-functions.R
\name{plot_N_uptake}
\alias{plot_N_uptake}
\title{Title}
\usage{
plot_N_uptake(conc = seq(0, 25, 0.25), spec_params, shape = "Michaelis-Menton")
}
\arguments{
\item{conc}{}

\item{spec_params}{}

\item{shape}{}
}
\value{
a ggplot object of nitrogen uptake rates
}
\description{
Title
}
\examples{
examples
}
