% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_lim.R
\name{Q_lim}
\alias{Q_lim}
\title{Q limitation on growth}
\usage{
Q_lim(Nf, Ns, spec_params)
}
\arguments{
\item{Nf}{Fixed nitrogen (mg m\eqn{^{-3}})}

\item{Ns}{Stored nitrogen (mg m\eqn{^{-3}})}

\item{spec_params}{A vector of named numbers. Must include:
\itemize{
\item \eqn{Q_{min}}, the minimum internal nutrient quotient
}}
}
\value{
a scalar of relative limitation from internal nutrient reserves on growth (between 0 and 1)
}
\description{
Calculates the internal nutrient quotient \eqn{Q} and its relative effect on growth rate \eqn{Q_{lim}} via:
\deqn{\begin{array}[ccc] 
Q &=& Q_{min} \left(1 + \frac{N_s}{N_f}\right)
Q_{lim} &=& \frac{Q - Q_{min}}{Q - K_c}
\end{array}}
}
\examples{
examples

}
