% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_rel.R
\name{Q_rel}
\alias{Q_rel}
\title{Relative nutrient quotient}
\usage{
Q_rel(Nf = NULL, Ns = NULL, Q_int = NULL, spec_params)
}
\arguments{
\item{Q_int}{Non-dimensionalised internal nutrient quotient}

\item{spec_params}{A vector of named numbers. Must include:
\itemize{
\item \code{Q_min}, the minimum internal nutrient quotient
\item \code{Q_max}, the maximum internal nutrient quotient
}}
}
\value{
the relative (0-1) internal nutrient quotient
}
\description{
Relative nutrient quotient
}
