% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroalgae-state-functions.R
\name{QQ}
\alias{QQ}
\title{Relative nutrient quotient}
\usage{
QQ(Q, spec_params)
}
\arguments{
\item{Q}{Non-dimentionalised internal nutrient quotient}

\item{spec_params}{A vector of named numbers. Must include:
\itemize{
\item \eqn{Q_{min}}, the minimum internal nutrient quotient
\item \eqn{Q_{max}}, the maximum internal nutrient quotient
\item \eqn{K_{c}}, the half-saturation constant for internal nutrient reserves
}}
}
\value{
the relative (0-1) internal nutrient quotient
}
\description{
Relative nutrient quotient
}
\examples{
examples
}
