% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_b.R
\name{u_b}
\alias{u_b}
\title{Relative water attenuation beneath canopy}
\usage{
u_b(
  U0,
  macro_state = c(biomass, hm),
  SA_WW = 0.5 * (0.0306/2),
  site_params,
  constants = c(s = 0.0045, gam = 1.13, a2 = 0.2^2, Cb = 0.0025)
)
}
\arguments{
\item{U0}{incoming incident water velocity (m/s)}

\item{macro_state}{vector of named numbers in the form c(biomass, hm) defining macroalgae wet weight (g) and height (m)}

\item{SA_WW}{conversion of wet weight to surface area (default is 0.5\emph{0.5}0.0306 for \textit{Macrocystis pyrifera})}

\item{site_params}{vector of named numbers in the form c(hz, hc, site_params\link{'d_top'}) defining total water depth (m), vertical water column occupied by the canopy (m), and depth of the top of the canopy beneath the water surface (m)}

\item{constants}{vector of named numbers in the form c(s = 0.0045, gam = 1.13, a2 = 0.2^2, Cb = 0.0025) defining extra constants for the attenuation submodel}
}
\value{
a scalar of relative water attenuation beneath canopy
}
\description{
Relative water attenuation beneath canopy
}
\examples{
examples
}
