% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_c.R
\name{u_c}
\alias{u_c}
\title{Relative water attenuation within canopy}
\usage{
u_c(
  U0,
  macro_state = c(biomass, hm),
  SA_WW = 0.5 * (0.0306/2),
  site_params,
  constants = c(s = 0.0045, gam = 1.13, a2 = 0.2^2, Cb = 0.0025)
)
}
\arguments{
\item{U0}{incoming incident water velocity (m/s)}

\item{macro_state}{vector of named numbers. Must include:
\itemize{
\item \code{biomass}, macroalgae wet weight (g)
\item \code{hm}, algae height (m)
}}

\item{SA_WW}{conversion of wet weight to surface area (default is 0.5\eqn{\times}0.5\eqn{\times}0.0306, based on \emph{Macrocystis pyrifera})}

\item{site_params}{vector of named numbers. Must include:
\itemize{
\item \code{hz}, total water depth (m)
\item \code{hc}, vertical water column occupied by the canopy (m)
\item \code{d_top}, depth of the top of the canopy beneath the water surface (m)
}}

\item{constants}{vector of named numbers defining extra constants for the attenuation submodel. Must include:
\itemize{
\item \code{s} = 0.0045
\item \code{gam} = 1.13
\item \code{a2} = 0.2^2
\item \code{Cb} = 0.0025
}}
}
\value{
the relative water attenuation coefficient (u_c)
}
\description{
#' @description
A short description...
}
\seealso{
\code{\link[=algae_height(), u_b(), C_t]{algae_height(), u_b(), C_t()}}
}
